requirements.txt
# ISTA Framework - MongoDB Edition
# Python package requirements

# MongoDB and Database
pymongo>=4.5.0              # MongoDB client library
dnspython>=2.4.0            # DNS support for MongoDB Atlas connection strings
motor>=3.3.0                # Async MongoDB driver (optional, for FastAPI)

# Data Generation
faker>=20.0.0               # Generate realistic fake data
factory-boy>=3.3.0          # Factory pattern support (optional)

# CLI and User Interface
click>=8.1.0                # Command-line interface creation
rich>=13.0.0                # Rich terminal output with colors/tables
typer>=0.9.0                # Alternative CLI framework (optional)

# Web and APIs
fastapi>=0.104.0            # FastAPI web framework (for data provisioning API)
uvicorn>=0.24.0             # ASGI server for FastAPI
starlette>=0.27.0           # Starlette dependency

# Data Validation and Serialization
pydantic>=2.0.0             # Data validation
pyyaml>=6.0                 # YAML parsing for data definitions
jsonschema>=4.19.0          # JSON schema validation

# Testing
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async pytest support
pytest-cov>=4.1.0           # Coverage reporting
pytest-xdist>=3.5.0         # Parallel test execution

# Database Tools
psycopg2-binary>=2.9.0      # PostgreSQL adapter (optional, for PostgreSQL support)
mysql-connector-python>=8.0.0  # MySQL adapter (optional, for MySQL support)
sqlalchemy>=2.0.0           # SQL toolkit (optional)

# Security and Secrets
python-dotenv>=1.0.0        # Load environment variables from .env
python-jose>=3.3.0          # JWT handling (optional, for auth)
passlib>=1.7.4              # Password hashing (optional)

# Logging and Monitoring
python-json-logger>=2.0.0   # Structured JSON logging (optional)
structlog>=23.0.0           # Structured logging (optional)

# Development Tools
black>=23.0.0               # Code formatter
flake8>=6.1.0               # Linter
pylint>=3.0.0               # Code analyzer
mypy>=1.7.0                 # Type checker
isort>=5.13.0               # Import sorter

# Documentation
sphinx>=7.0.0               # Documentation generator (optional)
sphinx-rtd-theme>=1.3.0     # ReadTheDocs theme (optional)

# Performance and Optimization
redis>=5.0.0                # Redis client (optional, for caching)
aiohttp>=3.9.0              # Async HTTP client (optional)

# Database-specific (uncomment as needed)
# psycopg2-binary>=2.9.0      # PostgreSQL
# mysql-connector-python>=8.0  # MySQL
# boto3>=1.28.0               # AWS/DynamoDB
